24/05

+++ src/main/java/pfe/example/demo/Service/ContributorService.java
@@ -8,6 +8,7 @@ import pfe.example.demo.Entites.*;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -16,18 +17,38 @@ public class ContributorService {
     ContributorRepository contributorRepository;
     @Autowired
     ListBlackRepository listBlackRepository;
+    @Autowired
+    AccountRepository accountRepository;
 
-    public void activateContributor(Long idContributeur) {
-        Contributor contributor = (Contributor) contributorRepository.getOne(idContributeur);
-        ListBlack listBlack = listBlackRepository.findByContributor(contributor);
-        this.listBlackRepository.delete(listBlack);
+
+    public void activateContributor(Long id) {
+        Optional<Contributor> contributor =  contributorRepository.findById(id);
+        if(contributor.isPresent()) {
+            Account account=contributor.get().getAccount();
+            if(account != null) {
+                account.setBlocked(false);
+                accountRepository.save(account);
+            }
+            ListBlack listBlack = listBlackRepository.findByContributor(contributor.get());
+            this.listBlackRepository.delete(listBlack);
+        }
     }
 
+
+
     public void blockedContributor(Long id) {
-        Contributor contributor = contributorRepository.getOne(id);
-        ListBlack listBlack = new ListBlack();
-        listBlack.setContributor(contributor);
-        listBlackRepository.save(listBlack);
+        Optional<Contributor> contributor =  contributorRepository.findById(id);
+
+        if(contributor.isPresent()) {
+            Account account=contributor.get().getAccount();
+            if(account != null) {
+                account.setBlocked(true);
+                accountRepository.save(account);
+            }
+            ListBlack listBlack = new ListBlack();
+            listBlack.setContributor(contributor.get());
+            listBlackRepository.save(listBlack);
+        }
     }
 
     public List<Contributor> findAll() {
